requirements.txt
requests>=2.31.0
reportlab==4.0.4
sentence-transformers==2.7.0
faiss-cpu==1.12.0
numpy>=1.24,<2.0


